+++ Commands/Misc/CmdTeleport.cs
@@ -26,7 +26,7 @@ namespace CommandDll
         public string Name { get { return "Teleport"; } }
         public CommandTypes Type { get { return CommandTypes.misc; } }
         public string Author { get { return "Gamemakergm"; } }
-        public int Version { get { return 1; } }
+        public decimal Version { get { return 1.00m; } }
         public string CUD { get { return ""; } }
         public byte Permission { get { return 80; } }
