+++ Commands/Moderation/CmdReview.cs
@@ -25,7 +25,7 @@ namespace CommandDll
         public string Name { get { return "Review"; } }
         public CommandTypes Type { get { return CommandTypes.mod; } }
         public string Author { get { return "Arrem"; } }
-        public int Version { get { return 1; } }
+        public decimal Version { get { return 1.00m; } }
         public string CUD { get { return ""; } }
         public byte Permission { get { return 0; } }
